Make maximum resident credential count configurable
Previously, we estimated that we can handle 100 resident keys when
returning the number of remaining resident keys in the credential
management command.

This patch introduces a config option to set a maximum count of resident
keys that is used to report the number of remaining resident keys and
that is enforced when trying to create a new resident key.