Clean up empty RP dirs if key store is full
If we try to register a new resident credential, we delete existing
credentials with the same RP and user ID.  If we then cannot store the
new credential because the credential limit is reached or there is a
filesystem write error, we may be left with an empty RP dir.  This patch
adds a check to delete empty RP dirs in this case.